Fixes file names and classes don't match.  All but PluginsSettings
