+++ src/modules/launcher/Wox.Infrastructure/Storage/IRepository.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 
 namespace Wox.Infrastructure.Storage
 {
+    [System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1649:File name should match first type name", Justification = "Generic, file is named correctly")]
     public interface IRepository<T>
     {
         void Add(T insertedItem);
