+++ src/modules/launcher/Wox.Infrastructure/Storage/ListRepository.cs
@@ -15,7 +15,8 @@ namespace Wox.Infrastructure.Storage
     /// The intent of this class is to provide a basic subset of 'list' like operations, without exposing callers to the internal representation
     /// of the data structure.  Currently this is implemented as a list for it's simplicity.
     /// </summary>
-    /// <typeparam name="T"></typeparam>
+    /// <typeparam name="T">typeof</typeparam>
+    [System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1649:File name should match first type name", Justification = "Generic, file is named correctly")]
     public class ListRepository<T> : IRepository<T>, IEnumerable<T>
     {
         public IList<T> Items { get { return _items.Values.ToList(); } }
