+++ src/modules/launcher/Wox.Infrastructure/UserSettings/PluginSettings.cs
@@ -39,18 +39,4 @@ namespace Wox.Infrastructure.UserSettings
             }
         }
     }
-
-    public class Plugin
-    {
-        public string ID { get; set; }
-
-        public string Name { get; set; }
-
-        public List<string> ActionKeywords { get; set; } // a reference of the action keywords from plugin manager
-
-        /// <summary>
-        /// Used only to save the state of the plugin in settings
-        /// </summary>
-        public bool Disabled { get; set; }
-    }
 }
